+++ mininterface/__init__.py
@@ -8,7 +8,7 @@ from .exceptions import Cancelled, InterfaceNotAvailable
 from .interfaces import get_interface
 
 from . import validators
-from .cli_parser import _parse_cli, assure_args
+from .cli_parser import parse_cli, assure_args
 from .subcommands import Command, SubcommandPlaceholder
 from .form_dict import DataClass, EnvClass
 from .mininterface import EnvClass, Mininterface
@@ -183,22 +183,22 @@ def run(env_or_list: Type[EnvClass] | list[Type[Command]] | None = None,
         start.choose_subcommand(env_or_list)
     elif env_or_list:
         # Load configuration from CLI and a config file
-        env, wrong_fields = _parse_cli(env_or_list, config_file, add_verbosity, ask_for_missing, args, **kwargs)
+        env, wrong_fields = parse_cli(env_or_list, config_file, add_verbosity, ask_for_missing, args, **kwargs)
     else:  # even though there is no configuration, yet we need to parse CLI for meta-commands like --help or --verbose
-        _parse_cli(_Empty, None, add_verbosity, ask_for_missing, args)
+        parse_cli(_Empty, None, add_verbosity, ask_for_missing, args)
 
     # Build the interface
-    interface = get_interface(title, interface, env)
+    m = get_interface(title, interface, env)
 
     # Empty CLI → GUI edit
     if ask_for_missing and wrong_fields:
         # Some fields must be set.
-        interface.form(wrong_fields)
-        {setattr(interface.env, k, v.val) for k, v in wrong_fields.items()}
+        m.form(wrong_fields)
+        {setattr(m.env, k, v.val) for k, v in wrong_fields.items()}
     elif ask_on_empty_cli and len(sys.argv) <= 1:
-        interface.form()
+        m.form()
 
-    return interface
+    return m
 
 
 __all__ = ["run", "Tag", "validators", "InterfaceNotAvailable", "Cancelled",