+++ mininterface/cli_parser.py
@@ -6,7 +6,7 @@ import sys
 import warnings
 from argparse import Action, ArgumentParser
 from contextlib import ExitStack
-from dataclasses import MISSING
+from dataclasses import MISSING, fields, is_dataclass
 from pathlib import Path
 from types import SimpleNamespace
 from typing import Optional, Sequence, Type, Union
@@ -15,12 +15,12 @@ from unittest.mock import patch
 import yaml
 from tyro import cli
 from tyro._argparse_formatter import TyroArgumentParser
-from tyro._fields import NonpropagatingMissingType
-# NOTE in the future versions of tyro, include that way:
-# from tyro._singleton import NonpropagatingMissingType
+from tyro._singleton import MISSING_NONPROP
 from tyro.extras import get_parser
 
-from .auxiliary import yield_annotations, yield_defaults
+
+from .auxiliary import yield_annotations
+from .config import MininterfaceConfig, Config
 from .form_dict import EnvClass, MissingTagValue
 from .tag import Tag
 from .tag_factory import tag_factory
@@ -137,8 +137,9 @@ def run_tyro_parser(env_or_list: Type[EnvClass] | list[Type[EnvClass]],
         with ExitStack() as stack:
             [stack.enter_context(p) for p in patches]  # apply just the chosen mocks
             res = cli(type_form, args=args, **kwargs)
-            if isinstance(res, NonpropagatingMissingType):
-                # NOTE tyro does not work if a required positional is missing tyro.cli() returns just NonpropagatingMissingType.
+            if res is MISSING_NONPROP:
+                # NOTE tyro does not work if a required positional is missing tyro.cli()
+                # returns just NonpropagatingMissingType (MISSING_NONPROP).
                 # If this is supported, I might set other attributes like required (date, time).
                 # Fail if missing:
                 #   files: Positional[list[Path]]
@@ -150,7 +151,7 @@ def run_tyro_parser(env_or_list: Type[EnvClass] | list[Type[EnvClass]],
         if ask_for_missing and getattr(e, "code", None) == 2 and eavesdrop:
             # Some required arguments are missing. Determine which.
             wf = {}
-            for arg in eavesdrop.partition(":")[2].strip().split(", "):
+            for arg in _fetch_eavesdrop_args():
                 treat_missing(type_form, kwargs, parser, wf, arg)
 
             # Second attempt to parse CLI
@@ -195,7 +196,6 @@ def treat_missing(env_class, kwargs: dict, parser: ArgumentParser, wf: dict, arg
         # However, the UI then is not able to use ex. the number filtering capabilities.
         # Putting there None is not a good idea as dataclass_to_tagdict fails if None is not allowed by the annotation.
         tag = wf[field_name] = tag_factory(MissingTagValue(),
-                                           # tag = wf[field_name] = tag_factory(MISSING,
                                            argument.help.replace("(required)", ""),
                                            validation=not_empty,
                                            _src_class=env_class,
@@ -205,17 +205,25 @@ def treat_missing(env_class, kwargs: dict, parser: ArgumentParser, wf: dict, arg
         # A None would be enough because Mininterface will ask for the missing values
         # promply, however, Pydantic model would fail.
         # As it serves only for tyro parsing and the field is marked wrong, the made up value is never used or seen.
-        if "default" not in kwargs:
-            kwargs["default"] = SimpleNamespace()
-        setattr(kwargs["default"], field_name, tag._make_default_value())
+        set_default(kwargs, field_name, tag._make_default_value())
 
 
-def _parse_cli(env_or_list: Type[EnvClass] | list[Type[EnvClass]],
-               config_file: Path | None = None,
-               add_verbosity=True,
-               ask_for_missing=True,
-               args=None,
-               **kwargs) -> tuple[EnvClass | None, dict, WrongFields]:
+def _fetch_eavesdrop_args():
+    return eavesdrop.partition(":")[2].strip().split(", ")
+
+
+def set_default(kwargs, field_name, val):
+    if "default" not in kwargs:
+        kwargs["default"] = SimpleNamespace()
+    setattr(kwargs["default"], field_name, val)
+
+
+def parse_cli(env_or_list: Type[EnvClass] | list[Type[EnvClass]],
+              config_file: Path | None = None,
+              add_verbosity=True,
+              ask_for_missing=True,
+              args=None,
+              **kwargs) -> tuple[EnvClass | None, dict, WrongFields]:
     """ Parse CLI arguments, possibly merged from a config file.
 
     Args:
@@ -228,45 +236,102 @@ def _parse_cli(env_or_list: Type[EnvClass] | list[Type[EnvClass]],
     Returns:
         Configuration namespace.
     """
+    if isinstance(env_or_list, list):
+        subcommands, env = env_or_list, None
+    else:
+        subcommands, env = None, env_or_list
+
     # Load config file
-    if config_file and isinstance(env_or_list, list):
-        # NOTE. Reading config files when using subcommands is not implemented.
-        static = {}
+    if config_file and subcommands:
+        # Reading config files when using subcommands is not implemented.
         kwargs["default"] = None
         warnings.warn(f"Config file {config_file} is ignored because subcommands are used."
-                      "It is not easy to set who this should work. "
-                      "Describe the developer your usecase so that they might implement this.")
-    if "default" not in kwargs and not isinstance(env_or_list, list):
+                      " It is not easy to set how this should work."
+                      " Describe the developer your usecase so that they might implement this.")
+
+    if "default" not in kwargs and not subcommands and config_file:
         # Undocumented feature. User put a namespace into kwargs["default"]
         # that already serves for defaults. We do not fetch defaults yet from a config file.
-        disk = {}
-        if config_file:
-            disk = yaml.safe_load(config_file.read_text()) or {}  # empty file is ok
-            # Nested dataclasses have to be properly initialized. YAML gave them as dicts only.
-            for key in (key for key, val in disk.items() if isinstance(val, dict)):
-                disk[key] = env_or_list.__annotations__[key](**disk[key])
-
-        # Fill default fields
-        if pydantic and issubclass(env_or_list, BaseModel):
-            # Unfortunately, pydantic needs to fill the default with the actual values,
-            # the default value takes the precedence over the hard coded one, even if missing.
-            static = {key: env_or_list.model_fields.get(key).default
-                      for ann in yield_annotations(env_or_list) for key in ann if not key.startswith("__") and not key in disk}
-            # static = {key: env_or_list.model_fields.get(key).default
-            #           for key, _ in iterate_attributes(env_or_list) if not key in disk}
-        elif attr and attr.has(env_or_list):
-            # Unfortunately, attrs needs to fill the default with the actual values,
-            # the default value takes the precedence over the hard coded one, even if missing.
-            # NOTE Might not work for inherited models.
-            static = {key: field.default
-                      for key, field in attr.fields_dict(env_or_list).items() if not key.startswith("__") and not key in disk}
-        else:
-            # To ensure the configuration file does not need to contain all keys, we have to fill in the missing ones.
-            # Otherwise, tyro will spawn warnings about missing fields.
-            static = {key: val
-                      for key, val in yield_defaults(env_or_list) if not key.startswith("__") and not key in disk}
-        kwargs["default"] = SimpleNamespace(**(disk | static))
+        disk = yaml.safe_load(config_file.read_text()) or {}  # empty file is ok
+        if mininterface := disk.pop("mininterface", None):
+            # Section 'mininterface' in the config file changes the global configuration.
+            for key, value in vars(_create_with_missing(MininterfaceConfig, mininterface)).items():
+                setattr(Config, key, value)
+        kwargs["default"] = _create_with_missing(env, disk)
 
     # Load configuration from CLI
-    env, wrong_fields = run_tyro_parser(env_or_list, kwargs, add_verbosity, ask_for_missing, args)
-    return env, wrong_fields
+    return run_tyro_parser(subcommands or env, kwargs, add_verbosity, ask_for_missing, args)
+
+
+def _create_with_missing(env, disk: dict):
+    """
+    Create a default instance of an Env object. This is due to provent tyro to spawn warnings about missing fields.
+    Nested dataclasses have to be properly initialized. YAML gave them as dicts only.
+    """
+
+    # Determine model
+    if pydantic and issubclass(env, BaseModel):
+        m = _process_pydantic
+    elif attr and attr.has(env):
+        m = _process_attr
+    else:  # dataclass
+        m = _process_dataclass
+
+    # Fill default fields with the config file values or leave the defaults.
+    # Unfortunately, we have to fill the defaults, we cannot leave them empty
+    # as the default value takes the precedence over the hard coded one, even if missing.
+    out = {}
+    for name, v in m(env, disk):
+        out[name] = v
+        disk.pop(name, None)
+
+    # Check for unknown fields
+    if disk:
+        warnings.warn(f"Unknown fields in the configuration file: {', '.join(disk)}")
+
+    # Safely initialize the model
+    return env(**out)
+
+
+def _process_pydantic(env, disk):
+    for name, f in env.model_fields.items():
+        if name in disk:
+            if isinstance(f.default, BaseModel):
+                v = _create_with_missing(f.default.__class__, disk[name])
+            else:
+                v = disk[name]
+        elif f.default is not None:
+            v = f.default
+        yield name, v
+
+
+def _process_attr(env, disk):
+    for f in attr.fields(env):
+        if f.name in disk:
+            if attr.has(f.default):
+                v = _create_with_missing(f.default.__class__, disk[f.name])
+            else:
+                v = disk[f.name]
+        elif f.default is not attr.NOTHING:
+            v = f.default
+        else:
+            v = MISSING_NONPROP
+        yield f.name, v
+
+
+def _process_dataclass(env, disk):
+    for f in fields(env):
+        if f.name.startswith("__"):
+            continue
+        elif f.name in disk:
+            if is_dataclass(f.type):
+                v = _create_with_missing(f.type, disk[f.name])
+            else:
+                v = disk[f.name]
+        elif f.default_factory is not MISSING:
+            v = f.default_factory()
+        elif f.default is not MISSING:
+            v = f.default
+        else:
+            v = MISSING_NONPROP
+        yield f.name, v